refactor: split changelog utilities
